requirements.txt
graphql-core~=0.4